RD-2004 Always expect csys-obj-type label to be present on Services
Due to a limitation of the filtering API, Services are expected to
always have a `csys-obj-type=service` label.

This is because previously, if services optionally have this label,
there is no way to filter out deployments that have
`csys-obj-type=service` OR `csys-obj-type is null`. OR operations are
not allowed in the filtering API.

Relevant backend changes to mark deployments as services by default will
be done in the near future.