RD-2004 Filter deployments shown in drill-down view (#1305)
* Show subdeployments of correct type in drill-down

Filter deployments shown in the drill-down view based on their type
(environments or services) based on the clicked drill-down button.

* Display an error message when using drilled-down widget wrong

An error message will be shown when using a drilled-down Deployments
View widget on a top-level page. This should guide the users and provide
them more context why it is not working and what they should do.

* Always expect csys-obj-type label to be present on Services

Due to a limitation of the filtering API, Services are expected to
always have a `csys-obj-type=service` label.

This is because previously, if services optionally have this label,
there is no way to filter out deployments that have
`csys-obj-type=service` OR `csys-obj-type is null`. OR operations are
not allowed in the filtering API.

Relevant backend changes to mark deployments as services by default will
be done in the near future.