RD-1527 Prevent clustering markers for selected deployments
Markers that correspond to selected deployments should always be
visible. This way the user receives feedback after changing the
deployment selection in the table, even when the deployment was
previously inside a cluster.

It also solves a technical problem that the leaflet.markercluster
library assumed that the CircleMarker that highlights the selected
marker is an element of its own and incremented the number of elements
in the cluster.