RD-1527 Deployments Map marker clustering (#1369)
* Install and use leaflet.markercluster
* Prevent clustering markers for selected deployments

Markers that correspond to selected deployments should always be
visible. This way the user receives feedback after changing the
deployment selection in the table, even when the deployment was
previously inside a cluster.

It also solves a technical problem that the leaflet.markercluster
library assumed that the CircleMarker that highlights the selected
marker is an element of its own and incremented the number of elements
in the cluster.

* Remove global L definition

It shadowed the `leaflet` module namespace which prevented accessing
types from plugins (e.g. `leaflet.markercluster`) which augmented the
main `leaflet` module.

This prevented consuming `L` as a global variable in ES modules. The
workaround is to access `L` from `window` (`window.L`)

* Maintain stable references for map marker positions

Use a `WeakMap` to maintain stable references to map marker positions.
Prevents the markercluster library from assuming that the markers moved
since the position object is new.

Fixes a UX bug with the cluster marker discarding its state and prevents
unnecessary computations.