RD-1532 Fix handling multiple note/warning/fix tags in README
When a widget README contained multiple note/warning/fix tags, the
transformation turned all the content between the first tag and the last
tag into a message, instead of transforming each tag into a message.

This is because the regexp used for matching those tags was greedy. It
has to be lazy (i.e. with a ? after *) to work correctly.

See http://www.regular-expressions.info/repeat.html#lazy