RD-1532 Update Deployments View widgets READMEs (#1388)
* Add README mappings for Deployments View widgets

Handle the regular and the drilled-down version of the widget.

* Show readmes for Deployments View widgets
* Update Deployments View widgets images
* Fix updateReadmes script to handle non-word character in titles

The script discarded non-word characters (e.g. parentheses, slashes)
when copying the titles from widgets documentation.

* Fix handling multiple note/warning/fix tags in README

When a widget README contained multiple note/warning/fix tags, the
transformation turned all the content between the first tag and the last
tag into a message, instead of transforming each tag into a message.

This is because the regexp used for matching those tags was greedy. It
has to be lazy (i.e. with a ? after *) to work correctly.

See http://www.regular-expressions.info/repeat.html#lazy