RD-1295 Move all TSC output to a tsc-dist directory
Do not pollute each individual project with its own dist.

The previous approach was problematic with the widgets project, because
having a dist directory in the widgets directory was treated by webpack
as another widget that should be copied and compiled. Having the tsc
output in a top-level directory solves the problem.