RD-2540 Fix race condition in filter_spec
There was a race condition between refetching the list of available
blueprints in the Resource Filter widget and the blueprint being
removed.

Removing the blueprint could have happened after the Resource Filter
widget's dropdown is opened, meaning that the test will fail as the
blueprint will supposedly still be in the system.