RD-1295 Add TS project references (#1404)
* Configure project references for backend and backend tests

Speeds up type-checking.
The transpiled JS files along with tsconfig.tsbuildinfo are stored in
backend/dist directory.

* Fix ESLint warnings and disable warning threshold

Any warning should fail linting.

* Add a solution tsconfig

A solution tsconfig includes all projects as references.

* Simplify type-checking scripts in package.json

Leave only a single check-types script because it is much simpler thanks
to using a solution tsconfig.

* Use a single ForkTsCheckerWebpackPlugin in build mode

Leverage project references to type-check the whole project during
development.

* Move all TSC output to a tsc-dist directory

Do not pollute each individual project with its own dist.

The previous approach was problematic with the widgets project, because
having a dist directory in the widgets directory was treated by webpack
as another widget that should be copied and compiled. Having the tsc
output in a top-level directory solves the problem.

* Create a solution file just for application projects

Create a tsconfig.ui.json solution file that has references to the
application project and the widgets project.
Use that solution file instead of the main tsconfig.json in webpack.
This way only relevant errors will be reported in webpack.
CI and the main `npm run check-types` still type-check all projects.