RD-410 Fix "Show related entities" icon placement
Fix the placement of the "Show related entities" icon of the task graph
in the Executions widget.

On Safari and other webkit engines, the icon appeared to be positioned
with respect to the root of the graph (the 0,0 position) instead of each
graph node.

Adding `display: inline` to the icon fixes the problem and it is
consistently positioned correctly across Chrome, Firefox, Epiphany (uses
webkit on Ubuntu) browsers.