RD-2005 Add subdeployment status icons in drill-down buttons (#1442)
* Add status icons in drill-down buttons

Show icons for the statuses of subdeployments in the drill-down butttons
in the details pane of the Deployments View widget.

Modify a test to check the correct icons are shown.

* Improve checking if an element contains a number

Add `cy.containsNumber` which makes sure that the expected number is a
standalone number and is not a substring of some other number.

For example, when checking `cy.contains('0')`, `10` would match even
though that is not the intention. `cy.containsNumber(0)` would fail for
`10` and only pass for `0`.