RD-2633 Add `setupIntercepts` callback to `mockLogin`
Allows setting up Cypress intercepts right before the page is loaded to
give the best chance of catching the right request.

Prevents problems where intercepts could be set up before the page from
the previous test is unloaded and a request from that previous page is
caught in the next test.