RD-1848 Add vertical slider in Deployments View (#1460)
* Use react-split-pane in Deployments View

Add a handle to change the width ratios of the table and the details
pane in the Deployments View widget.

* Transform Grid layout to TypeScript
* Move `useWidthObserver` to `Stage.Hooks`

* Use useWidthObserver instead of WidthProvider for layout's width

`WidthProvider` from `react-grid-layout` only detects width changes when
the browser resizes. Other changes, e.g. the component nearby changing
its size, do not cause the grid layout to change its width, causing
widgets' size to be outdated and not matching the container.

`useWidthObserver` fixes that by listening to any changes in the width,
whether due to browser resizing or other components nearby.

* RD-1848 Add a test for resizing the master-details view