RD-2593 Hide "Edit a copy in Composer" button (#1466) (#1475)
* Add a configuration flag to hide "Edit in Composer" button

Add a configuration option to Blueprint Action Buttons widget to show
the "Edit a copy in Composer" button. The flag is disabled by default.

* Add a test for visibility of Edit in Composer button

Refactor the tests structure for Blueprint Action Buttons widget.
Add a test to verify that the "Edit a copy in Composer" button is not
visible by default and becomes visible after changing the configuration.