RD-2747 Fix connecting Terraform dependencies in Topology
It can happen that a Terraform dependency will have 3 dot-separated
parts, with the last one being the actual dependency name. The logic
needs to always take the last part, instead of taking the second part
(which was the last one when there were just 2 dot-separated parts in
the dependency name).