RD-1827 Remove double loader in DeploymentActionButtons widget (#1506)
* Improve data fetching in DeploymentActionButtons widget

Have separate states for loading, error, success.
Log an error to the console when there is no deployment selected.
Do not show a second spinner when fetching workflows.

* Fix static propTypes on PagesList

propTypes and defaultProps were declared as instance properties, not
static properties. It caused errors to be logged by React.

* Allow using Error as WidgetData

Errors have 0 own properties. It caused them to be miscategorized as
empty widget data.