RD-2747 Fix expanding Terraform nodes in Topology widget (#1501)
* Fix connecting Terraform dependencies in Topology

It can happen that a Terraform dependency will have 3 dot-separated
parts, with the last one being the actual dependency name. The logic
needs to always take the last part, instead of taking the second part
(which was the last one when there were just 2 dot-separated parts in
the dependency name).

* Add some type annotations to Topology

The Topology component is not migrated to TS yet. This commit adds some
defensive type annotations, but does not fully migrate the component to
TS.