RD-1851 Expose WidgetlessToolbox on Stage.Types
Describes a Toolbox which was created without a specific component.

It is useful in components in the widget's configuration - such
components can require access to the Manager and need to use other UI
components that require `Toolbox`, but they don't need the full widget's
`Toolbox`, only some specific widget-agnostic methods.