RD-1851 Adjust test for filter ID config in Deployments View
Adjust the test for the filter ID configuration in the Deployments View
widget. It now verifies the autocomplete functionality of the added
component.

The component should allow clearing the existing filter, as well as
setting it to a valid filter ID. If the user types in an invalid filter
ID, it should not allow submitting the configuration.