RD-1851 Add autocomplete for default filter in Deployments View widget (#1520)
* Add autocomplete for filter ID configuration in Deployments View

Use a DynamicDropdown as a custom component in the configuration.
Leverage WidgetlessToolbox instead of the regular widget-specific
Toolbox.

* Adjust test for filter ID config in Deployments View

Adjust the test for the filter ID configuration in the Deployments View
widget. It now verifies the autocomplete functionality of the added
component.

The component should allow clearing the existing filter, as well as
setting it to a valid filter ID. If the user types in an invalid filter
ID, it should not allow submitting the configuration.

* Add a test to verify Deployments View with an invalid filter ID

Add a test to make sure that the Deployments View widget handles a case
when the filter ID set in the widget configuration is no longer valid.

This can happen when the filter used in the widget is removed and the
widget keeps its configuration.