RD-2790 Change the structure of drill-down buttons
Change the structure of drill-down buttons to avoid showing the loading
overlay on top of the "Details" button, which does not require any data
fetching.

Presently, the "Details" button will always be enabled, and the loading
overlay will appear only on top of the subdeployment drilldown buttons.

This improves the UX.